Moved edit profile stuff to api + fixed it up
