+++ mobile-app/src/components/HeaderOptions.tsx
@@ -17,6 +17,7 @@ import { auth, firestore } from "../firebase/config";
 import { collection, where } from "@firebase/firestore";
 import { doc, getDocs, query, setDoc } from "firebase/firestore";
 import { userT, userWithIdT } from "../lib/db_types";
+import { updateProfileDataInDB } from "../firebase/api";
 
 function sharedOptions(colorScheme: string): NativeStackNavigationOptions {
   return {
@@ -220,49 +221,7 @@ export function editProfileOptions(
       <RoundedButton
         text="Done"
         icon={IconCheck}
-        onPress={async () => {
-
-          
-
-          // if (auth.currentUser == null) {
-          //   console.log("Cannot edit profile: User not logged in")
-          //   return;
-          // }
-
-
-          //TODO Ensure username is unique
-          const usersRef = collection(firestore, "users");
-          const q = query(usersRef, where("username", "==", profileFormData.username))
-          const queryResults = await getDocs(q);
-
-          // Check if were changing username, if so check if username already exist in DB
-          if (userData.username !== profileFormData.username && queryResults.size >= 1) {
-            console.log("Error: Username already taken"); // Display this on screen instead
-            profileFormData.username = "";
-            return;
-          }
-
-          // On success
-          const newDataForAtom: userWithIdT = {
-            ...userData,
-            ...profileFormData,
-          }
-
-          const newDataForDb: userT = {
-            username: profileFormData.username,
-            displayName: profileFormData.displayName,
-            picture: userData.picture,
-            createdAt: userData.createdAt
-          }
-
-          //Push Data to DB
-          const userDocRef = doc(firestore, "users", userData.id) //change to auth.currentUser.uid when auth is fixed
-          await setDoc(userDocRef, newDataForDb)
-
-          //Update the current atoms accordingly 
-          setUserData(newDataForAtom)
-          router.back();
-        }}
+        onPress={() => updateProfileDataInDB(profileFormDataAtom,currentUserAtom)}
       />
     ),
     headerTitleAlign: "center",
